requirements.txt
Django~=4.2
openai~=0.27.4
whitenoise~=6.2.0
python-dotenv~=0.20.0
djangorestframework~=3.14.0
opencv-python~=4.7.0.72
numpy~=1.24.2
Pillow~=9.5.0
mediapipe~=0.9.2.1
django-q==1.3.9
django-jazzmin==2.6.0